move WidgetRoute to mod
