+++ src/pages/mod.rs
@@ -4,22 +4,19 @@ use dioxus_router::prelude::*;
 use strum::IntoEnumIterator;
 use strum_macros::EnumIter;
 
-pub mod route_trait;
-pub mod media;
-pub mod generator;
 pub mod converter;
+pub mod encoder_decoder;
+pub mod generator;
 pub mod home_page;
 pub mod layout;
-pub mod encoder_decoder;
+pub mod media;
 
-use home_page::HomePage;
-use layout::{Container, WidgetView};
-use generator::GeneratorRoute;
 use converter::ConverterRoute;
 use encoder_decoder::EncoderDecoderRoute;
+use generator::GeneratorRoute;
+use home_page::HomePage;
+use layout::{Container, WidgetView};
 use media::MediaRoute;
-
-use self::route_trait::WidgetRoute;
 
 #[rustfmt::skip]
 #[derive(Clone, Debug, EnumIter, PartialEq, Routable)]
@@ -95,6 +92,32 @@ impl Route {
             _ => vec![],
         }
     }
+}
+
+pub trait WidgetRoute: Routable + IntoEnumIterator + PartialEq + Clone {
+    fn get_widget_routes() -> Vec<Route>;
+
+    fn get_widgets() -> Vec<Self> {
+        Self::iter()
+            .filter(|route| route.get_widget_entry().is_some())
+            .collect()
+    }
+
+    fn get_widget_title_string(&self) -> Option<&'static str> {
+        Some(self.get_widget_entry()?.title)
+    }
+
+    fn get_widget_short_title_string(&self) -> Option<&'static str> {
+        Some(self.get_widget_entry()?.short_title)
+    }
+
+    fn get_widget_description_string(&self) -> Option<&'static str> {
+        Some(self.get_widget_entry()?.description)
+    }
+
+    fn get_widget_type_string() -> &'static str;
+
+    fn get_widget_entry(&self) -> Option<&'static WidgetEntry>;
 }
 
 #[derive(PartialEq, Eq, Clone, Copy, Debug)]
